Remove BaseLineMarkerTest
