+++ src/test/kotlin/com/daveme/chocolateCakePHP/test/BaseTestCase.kt
@@ -1,6 +1,13 @@
 package com.daveme.chocolateCakePHP.test
 
+import com.intellij.codeInsight.daemon.LineMarkerInfo
+import com.intellij.codeInsight.daemon.LineMarkerProvider
+import com.intellij.codeInsight.daemon.LineMarkerProviders
+import com.intellij.codeInsight.daemon.RelatedItemLineMarkerInfo
+import com.intellij.navigation.GotoRelatedItem
+import com.intellij.psi.PsiElement
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import kotlin.reflect.KClass
 
 public abstract class BaseTestCase : BasePlatformTestCase() {
 
@@ -8,4 +15,21 @@ public abstract class BaseTestCase : BasePlatformTestCase() {
         return "src/test/fixtures"
     }
 
+    fun calculateLineMarkers(element: PsiElement, klass: KClass<out LineMarkerProvider>): List<LineMarkerInfo<*>> {
+        return LineMarkerProviders.getInstance().allForLanguage(element.language)
+            .flatMap { provider ->
+                if (provider::class != klass) {
+                    return@flatMap listOf<LineMarkerInfo<*>>()
+                }
+                val markers = arrayListOf<LineMarkerInfo<*>>()
+                provider.collectSlowLineMarkers(listOf(element), markers)
+                return@flatMap markers
+            }
+    }
+
+    protected fun gotoRelatedItems(marker: LineMarkerInfo<*>): MutableCollection<out GotoRelatedItem> {
+        val renderer = marker.createGutterRenderer() as LineMarkerInfo.LineMarkerGutterIconRenderer<*>
+        val lineMarkerInfo = renderer.lineMarkerInfo as RelatedItemLineMarkerInfo<*>
+        return lineMarkerInfo.createGotoRelatedItems()
+    }
 }
